apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
celery[redis]==5.4.0
redis==5.0.4
pdfplumber==0.11.0
pytesseract==0.3.10
Pillow==10.3.0
python-multipart==0.0.9
thefuzz[speedup]==0.22.1
openai==1.51.0
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
